GUI: (QIodeAbstractTabWidget) Abstract class to be used above QIodeTabWidget so that we can connect databaseModified signal from AbstractTableView to the tabDatabaseModified slot
