+++ gui/CMakeLists.txt
@@ -47,6 +47,7 @@ set(GUI_CUSTOM_WIDGETS
     custom_widgets/text_editor.h                custom_widgets/text_editor.cpp
     custom_widgets/report_editor.h              custom_widgets/report_editor.cpp
     custom_widgets/file_delegate.h              custom_widgets/file_delegate.cpp
+    custom_widgets/tab_widget_abstract.h        custom_widgets/tab_widget_abstract.cpp
     custom_widgets/tab_widget.h                 custom_widgets/tab_widget.cpp
     custom_widgets/file_explorer_proxy.h
     custom_widgets/file_explorer.h              custom_widgets/file_explorer.cpp
